requirements.txt
django-environ==0.8.1
greenlet==1.1.2
PyMySQL==1.0.2
PyYAML==6.0
SQLAlchemy==1.4.32
